% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{bootstrapper}
\alias{bootstrapper}
\title{Bootstrapped variance estimates}
\usage{
bootstrapper(
  LL,
  x,
  w = 1,
  noParams,
  method = "Local",
  loops = 600,
  alpha,
  sigma,
  gamma = NULL,
  nu = NULL
)
}
\arguments{
\item{LL}{A function that specifies the (negative) log likelihood derived from the pdf. Parameters are named to make problem easier. Values are named x and weights are named w.}

\item{x}{Data values used in parameter estimation.}

\item{w}{Data weights used in estimation. Default = 1.}

\item{noParams}{The number of parameters being optimised over in LL (either alpha or alpha and sigma.}

\item{method}{The algorithm used in the search function. If "Global", uses SANN in optim function. If "Local", uses L-BFGS-B in optim function. Default is local.}

\item{loops}{Number of loops used to construct the distribution the variances are calculated from. Default is 600.}

\item{alpha}{Value of alpha parameter in pareto distribution.}

\item{sigma}{Value of sigma parameter in pareto distribution.}

\item{gamma}{Value of gamma parameter. If NULL, set to lower threshold or minimum of x.}

\item{nu}{Value of gamma parameter. If NULL, set to upper threshold or maximum of x.}
}
\value{
Returns matrix of mean, variance, bias, and bias significance for alpha, sigma, gamma, and nu.
}
\description{
Calculates confidence bounds using repeated estimates of likelihood function with bootstrapped draws from dataset.
}
\examples{
\dontrun{
LL <- function(param, alpha, sigma){
alpha <- param[1]
sigma <- param[2]
logl <-  sum(w(log(1/sigma) - ((1+alpha)/alpha)log((sigma +alpha(x-gamma))/sigma)))
return(-logl)
}
bootstrapper(LL, x=Value, w=Weights, noParams=2, method="Local", loops=600, alpha=1.8, sigma=1.2, gamma=min(Value), nu=max(Value))
}
}
