% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateSample.R
\name{calibrateSample}
\alias{calibrateSample}
\title{Create samples from estimated top tail using calibrate approach.}
\usage{
calibrateSample(
  x,
  w = 1,
  id = NA,
  inputDemoName = NULL,
  inputDemo = inputDemo,
  alpha = 1.8,
  gamma = NULL,
  sigma = NA,
  specification = "Type 1",
  loopmax = 2000
)
}
\arguments{
\item{x}{Data values used in parameter estimation.}

\item{w}{Data weights used in estimation. Default = 1.}

\item{id}{ID values for observations from database to help with linking. Default is NA.}

\item{inputDemoName}{Names for variables in inputDemo. If NULL (default), set names to column names of inputDemo.}

\item{inputDemo}{Matrix (nxm) of demographic variables of households used to calibrate weights. Must be same length as x and w.}

\item{alpha}{Value of alpha parameter in pareto distribution.}

\item{gamma}{Value of gamma parameter. If NULL, set to lower threshold or minimum of x.}

\item{sigma}{Value of sigma parameter in generalised pareto distribution.}

\item{specification}{Defines distribution CDF used for estimating population. Can be "Type 1" (default)  or "Generalized".}

\item{loopmax}{Maximum number of iterations. Default is 2000.}
}
\value{
Returns data frame of x, w, and id adjusted to match pareto and calibration criteria.
}
\description{
Generate new observations from a specified distribution using the randraw sampling approach in Cantarella, Neri & Ranalli, 2021.
}
\examples{
\dontrun{
calibrateSample(x=Value,w=Weight,id=NA,inputDemoName=c("Gender","Age","Employment"),inputDemo=data[,c("Gender","Age","Employment")],alpha=1.8,gamma=NULL,sigma=NA,specification="Type 1",loopmax=2000)
}
}
