% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatedParetoPopFunction.R
\name{estimatedParetoPopFunction}
\alias{estimatedParetoPopFunction}
\title{Estimate total pareto population}
\usage{
estimatedParetoPopFunction(
  x,
  w,
  alpha,
  gamma = NULL,
  sigma = NA,
  specification = "Type 1"
)
}
\arguments{
\item{x}{Data values used in parameter estimation.}

\item{w}{Data weights used in estimation. Default = 1.}

\item{alpha}{Value of alpha parameter in pareto distribution.}

\item{gamma}{Value of gamma parameter. If NULL, set to lower threshold or minimum of x.}

\item{sigma}{Value of sigma parameter in generalised pareto distribution.}

\item{specification}{Defines distribution CDF used for estimating population. Can be "Type 1" (default)  or "Generalized".}
}
\value{
Returns a table of total population in pareto tail (including origianl values and the newly estimated m2 population), the total population of the tail (popu), and the number of added points (Lm2).
}
\description{
Function extrapolates the weights in proportion of CDF covered by data (m1= sum(weights)) to estimate the total population (m=m1/F(xmax)) and population of uncovered area (m2).
}
\examples{
\dontrun{
estimatedParetoPopFunction(x=Value,w=Weights,alpha=1.8,gamma=min(Value),sigma=1.2,specification="Generalized")
}
}
