% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HillLoopfunction.R
\name{HillLoopfunction}
\alias{HillLoopfunction}
\title{Hill parameter estimator}
\usage{
HillLoopfunction(x, w)
}
\arguments{
\item{x}{Data values used in parameter estimation.}

\item{w}{Data weights used in estimation. Default = 1.}
}
\value{
Length of observation of x that should be used to combat bias in pareto Type 1 estimate.
}
\description{
Loops over dataset to find Hill parameter r that maximises the fit to the Pareto Type 1 distribution.
}
\examples{
\dontrun{
HillLoopfunction(Value,Weights)
}
}
