% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HardSandwichMaker.R
\name{HSandwichMaker}
\alias{HSandwichMaker}
\title{Hardcoded confidence bounds for pseudo-maximum likelihood}
\usage{
HSandwichMaker(
  x,
  w = 1,
  gamma = NULL,
  nu = NULL,
  specification,
  truncated,
  ParamNames,
  ParamEstimates
)
}
\arguments{
\item{x}{Data values used in parameter estimation.}

\item{w}{Data weights used in estimation. Default = 1.}

\item{gamma}{Lower threshold parameter for pareto. Default is minimum of x.}

\item{nu}{Upper threshold parameter for pareto. Default is maximum of x.}

\item{specification}{Evaluate sandwich estimate for "Type 1" or "Generalized" (hard coded because of calling gamma).}

\item{truncated}{Evaluate sandwich estimate for truncated (TRUE) or untruncated (FALSE) (hard coded because of calling gamma).}

\item{ParamNames}{The names of the parameters used in the expression. Example for generalised pareto: c("alpha", "sigma")}

\item{ParamEstimates}{Estimates obtained for the parameters. Following ordering of ParamNames}
}
\value{
Returns confidence bands for each of the variables, and the variance-covariance matrix
}
\description{
Calculates confidence bounds for pseudo-maximum likelihood using sandwich variance-covariance matrix (Geyer (2013): https://www.stat.umn.edu/geyer/5601/notes/sand.pdf).
}
\examples{
\dontrun{
HSandwichMaker(x=Value, w=Weights,gamma=100,nu=max(Value) specification="Generalized", truncated=FALSE, ParamNames=c("alpha", "sigma"),ParamEstimates=c(1.8,1.2))
}
}
