% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPara.R
\name{modelPara}
\alias{modelPara}
\title{Model selection code}
\usage{
modelPara(output.estimates, model.selection = "BIC", ksp = 0.01, ttp = 0.1)
}
\arguments{
\item{output.estimates}{The table output from the pareto code providing parameter value estimates, test values, and model-selection criterion}

\item{model.selection}{The method used to select the model from the arms approach (see Zwijnenburg, Grilli & Engelbrecht, 2022), akaike information criterion, bayesian information criterion, or log likelihood. Default = "BIC".}

\item{ksp}{The level of significant p-value used to test fit of data to proposed distribution. Default is 0.01.}

\item{ttp}{The level of significant p-value used to test for evidence of finite upper truncation point. Default is 0.1.}
}
\value{
Returns the selected specification (from "Type 1" or "Generalized") and if the distribution has an upper truncation value (TRUE or FALSE)
}
\description{
Part of the centralised approach code, taking the output table to identify the model specification and whether there is upper truncation.
}
\examples{
\dontrun{
modelPara(output.estimates=output.estimates,model.selection="Arms",ksp=0.01,ttp=0.1)
}
}
