% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parame.R
\name{parame}
\alias{parame}
\title{Model parameter value code}
\usage{
parame(
  output.estimates,
  specification = "Type 1",
  top.trunc = TRUE,
  calibrate.ALPHA = NULL,
  calibrate.SIGMA = NULL
)
}
\arguments{
\item{output.estimates}{The table output from the pareto code providing parameter value estimates, test values, and model-selection criterion}

\item{specification}{The model specification parameters from output.estimates will be taken from ("Type 1" or "Generalized").}

\item{top.trunc}{The inclusion of the upper truncation parameter from output.estimates (TRUE or FALSE).}

\item{calibrate.ALPHA}{Used to set alpha parameter if not using results. Default is NULL.}

\item{calibrate.SIGMA}{Used to set sigma parameter if not using results. Default is NULL.}
}
\value{
Returns the preferred specification parameter estimates for alpha, gamma, sigma, and nu.
}
\description{
Part of the centralised approach code, returns parameter values based on estimates and selected model form.
}
\examples{
\dontrun{
parame(output.estimates=output.estimates,specification="Type 1",top.trunc=TRUE,calibrate.ALPHA=NULL,calibrate.SIGMA=NULL)
}
}
