% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParetoAnalysis.R
\name{ParetoAnalysis}
\alias{ParetoAnalysis}
\title{Runs centralised apprach (Zwijnenburg, Grilli & Engelbrecht, 2022) Pareto code}
\usage{
ParetoAnalysis(
  inputValues = NULL,
  inputWeights = NULL,
  inputid = NULL,
  inputDemo = NULL,
  inputDemoName = NULL,
  x_0 = NULL,
  type,
  trun = "Untruncated",
  method = "RanDraw",
  simnum = 1000,
  pre.specification = NULL,
  pre.top.trunc = NULL,
  graphs = FALSE,
  ksp = 0.01,
  ttp = 0.1,
  upperT = 0.85,
  lowerT = 0.15,
  sim.ALPHA = 1.8,
  sim.SIGMA = 1.2,
  loopmax = 2000,
  calibrate.ALPHA = NULL,
  calibrate.SIGMA = NULL,
  HillEstimator = NULL,
  OptimSearch = "Default",
  seeded = NULL,
  Sampling = "Deterministic",
  model.selection = c("Arms", "Loglikelihood", "AIC", "BIC"),
  simulatePopulation = TRUE
)
}
\arguments{
\item{inputValues}{Vector (nx1) of values held by household}

\item{x}{Data values used in parameter estimation.}

\item{inputWeights:}{Vector (nx1) of weights of households}

\item{inputid:}{Vector (nx1) of ids of households}

\item{inputDemo:}{Matrix (nxm) of demographic variables of households}

\item{inputDemoName:}{Vector(1xm) of names of demographic variables}

\item{x_0:}{Threshold value (numeric) or strings "top10", "top5", "top1" to select top tail of input data. Any string input takes form "top##", with the number ## being extracted.}

\item{type:}{Simulation Option for Pareto Type 1 ("Type 1"), Generalized Pareto ("Generalized"), or EU-SILC ("eusilc") data}

\item{trun:}{Simulation Option for Untruncated ("Untruncated"), Truncated ("Truncated"), or randomly truncated ("randTruncated") data.}

\item{method:}{Draws from estimated Pareto distribution using added rich households ("Synths"), adjusted weights ("Calibrate"), or adjusted values ("ranDraw")}

\item{simnum:}{Simulation Option for number of observations to be drawn from Pareto Type 1 or Generalized Pareto distributions}

\item{pre.specification:}{Adjustment Option to overwrite goodness-of-fit results and select distribution type ("Type 1" or "Generalized")}

\item{pre.top.trunc:}{Adjustment Option to overwrite truncation test and select if truncated (TRUE or FALSE)}

\item{graphs:}{TRUE/FALSE option to produce graphs}

\item{ksp:}{Kolomogorov-Smirnov test statistic p-value level of significance to reject the null hypothesis that the data is drawn from the distribution}

\item{ttp:}{Truncation test statistics p-value level of significance to reject null hypothesis that the distribution is not truncated}

\item{upperT:}{Option for simulated Truncated ("Truncated") data to vary the percentage truncated at top of the distribution}

\item{lowerT:}{Option for simulated Truncated ("Truncated") data to vary the percentage truncated at bottom of the distribution}

\item{sim.ALPHA:}{Option for simulated  data to vary the Pareto Shape Parameter}

\item{sim.SIGMA:}{Option for simulated  data to vary the Pareto Scale Parameter}

\item{loopmax:}{Maximum number of iterations in the Calibrate iterative step}

\item{calibrate.ALPHA:}{Option for setting Pareto Shape Parameter when using pre.specification}

\item{calibrate.SIGMA:}{Option for setting Pareto Scale Parameter when using pre.specification}

\item{HillEstimator:}{Option for setting Hill Estimator share \link{0,1} of data used in estimating shape parameters. If NULL, it is calculated as either 0.5, or optimised if sample is small enough. Default 1 (i.e. no Hill Estimator adjustment)}

\item{OptimSearch:}{Sets the search style used for likelihood optimisation. "Default" uses existing options, or can set all to "Global", "Local", or "Both" (runs global then local).}

\item{seeded:}{Sets the seed to have fixed random draws (default: 20200916). Can therefore also use to alter seed. If set to NULL, then seed is not set (so can be randomised outside of the code, or doesnt overwrite existing seed setting.)}

\item{Sampling:}{Method of drawing new observations in the tail. Default option is "Deterministic" (calculate value from CDF where next observation would occur). Alternative is "Inverse" (make m2 random draws in the uncovered region of the tail)}

\item{model.selection:}{Method for choosing the winning model. Default option is "Arms", which is a logical decision tree to choose the smallest model, favouring trunction, from those retained in KS test.}

\item{simulatePopulation:}{If TRUE (default), households are simulated in the estimated distribution. If FALSE, code ends with Pareto estimates output.}
}
\value{
Returns estimated values, variances, goodness of fit and information criterion for pareto model. If simulatePopulation=TRUE, also returns estimated data sampled from pareto using method. If graphs=TRUE, also provides test graphical output.
}
\description{
Estimates pareto values from input data (or simulated data) with option to simulate data from estimated pareto using multiple approaches.
}
\examples{
ParetoAnalysis(inputValues=Value,inputWeights=Weight,id=hid)
ParetoAnalysis(type="Type 1",trun = "Truncated")
ParetoAnalysis(x_0 = 100, type="Type 1",trun = "Truncated", simnum=10000, upperT = 0.85, lowerT=0.15,sim.ALPHA=1.8,sim.SIGMA=1.2)
ParetoAnalysis(inputValues=Value,inputWeights=Weight,id=hid, pre.specification="Generalized", pre.top.trunc="TRUE", calibrate.ALPHA=1.8,calibrate.SIGMA=1.2)
ParetoAnalysis(inputValues=Value,inputWeights=Weight,id=hid,x_0="top5",method="Synths", graphs=TRUE, ksp=0.05, ttp = 0.05, HillEstimator=1, OptimSearch="Local" Sampling="Inverse", model.selection = "BIC", simulatePopulation = FALSE)
}
