% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pfunction.R
\name{Pfunction}
\alias{Pfunction}
\title{Analytical solution to pareto type 1 parameters}
\usage{
Pfunction(x, w = 1, r = NULL)
}
\arguments{
\item{x}{Data values used in parameter estimation.}

\item{w}{Data weights used in estimation. Default = 1.}

\item{r}{The breakpoint for the Hill estimator. Must be between 1 and n-1. If NULL (default), r=length(x)-1.}
}
\value{
Returns parameter estimates for alpha and sigma (and the minimum and maximum for x) for the truncated generalised pareto.
}
\description{
Returns parameters following the type 1 pareto form, using analytical solution from Aban, Meerschaert & Panorska (2006).
}
\examples{
\dontrun{
Pfunction(Value,w=Weights,r=length(Value)-1)
}
}
