% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPlots.R
\name{PPlots}
\alias{PPlots}
\title{Zipf and Kolmogorov-Smirnov graphical tests for Pareto fit}
\usage{
PPlots(
  ValueEquation,
  RankEquation,
  TypeName = NA,
  inputData_Values,
  inputData_Weights = 1,
  ksp = 0.01,
  graphs = TRUE
)
}
\arguments{
\item{ValueEquation}{Single variable function defining the inverse CDF function for the distribution of interest. Used to sample data. Example for Pareto Type 1: function(x){(gamma.pt1.hat2*(1-x)^(-1/alpha.pt1.hat2))}.}

\item{RankEquation}{Single variable function defining the complementary cumulative distribution function (CCDF). Used to rank x from 0 to 1.  Example for Pareto Type 1: function(x){(gamma.pt1.hat2/x)^alpha.pt1.hat2}.}

\item{TypeName}{Name of distribution.}

\item{inputData_Values}{Data values you wish check to match distribution used for RankEquation.}

\item{inputData_Weights}{Weights attached to data points. Default = 1.}

\item{ksp}{The level of significant p-value used to test fit of data to proposed distribution. Default is 0.01.}

\item{graphs}{If TRUE, graphs of tests are generated.}
}
\value{
Returns the Kolmogorov-Smirnov test p-value. When graphs=TRUE, also returns Zipf, Kolmogorov-Smirnov, and residual plots.
}
\description{
Function tests the fit of input data to a specified distribution.
}
\examples{
PPlots(function(x){(gamma.pt1.hat2*(1-x)^(-1/alpha.pt1.hat2))},function(x){(gamma.pt1.hat2/x)^alpha.pt1.hat2},"Pareto Type 1",simu$Value, simu$Weights,ksp,TRUE)
}
