% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randrawSample.R
\name{randrawSample}
\alias{randrawSample}
\title{Create samples from estimated top tail using randraw sampling.}
\usage{
randrawSample(
  x,
  w = 1,
  id = NA,
  m2 = NULL,
  popu = NULL,
  alpha = 1.8,
  gamma = 100,
  sigma = NA,
  Sampling = "Inverse",
  specification = "Type 1"
)
}
\arguments{
\item{x}{Data values used in parameter estimation.}

\item{w}{Data weights used in estimation. Default = 1.}

\item{id}{ID values for observations from database to help with linking. Default is NA.}

\item{m2}{Estimated number of draws required to complete pareto distribution. Output from estimatedParetoPopFunction function. If NULL (default), is calculated using parameter values, specified form and maximum(x), following approach in estimatedParetoPopFunction.}

\item{popu}{Estimated total number of observations in pareto distribution. Output from estimatedParetoPopFunction function. If NULL (default), is calculated using parameter values, specified form and maximum(x), following approach in estimatedParetoPopFunction.}

\item{alpha}{Value of alpha parameter in pareto distribution.}

\item{gamma}{Value of gamma parameter. If NULL, set to lower threshold or minimum of x.}

\item{sigma}{Value of sigma parameter in generalised pareto distribution.}

\item{Sampling}{Method of drawing values from the uncovered region. Can be "Deterministic" (method in Zwijnenburg, Grilli & Engelbrecht, 2022) or "Inverse" for inverse sampling.}

\item{specification}{Defines distribution CDF used for estimating population. Can be "Type 1" (default)  or "Generalized".}
}
\value{
Returns data frame with original values, weights, and the new values calculated for each observation.
}
\description{
Generate new observations from a specified distribution using the randraw sampling approach in Zwijnenburg, Grilli & Engelbrecht, 2022.
}
\examples{
specification(x=Value,w=Weights,id=hid,m2=m2,popu=popu,alpha=1.8,gamma=100,sigma=NA,Sampling="Inverse",specification="Type 1")
}
