% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandwichMaker.R
\name{SandwichMaker}
\alias{SandwichMaker}
\title{Calculates confidence bounds for pseudo-maximum likelihood}
\usage{
SandwichMaker(x, w = 1, EXPRESS, ParamNames, ParamEstimates)
}
\arguments{
\item{x}{Data values used in parameter estimation.}

\item{w}{Data weights used in estimation. Default = 1.}

\item{EXPRESS}{expression function giving the log likelihood function. This is differentiated with respect to ParamNames. Example for generalised pareto: expression(w(log(1/sigma) - ((1+alpha)/alpha)log((sigma +alpha(x-100.0129))/sigma)))}

\item{ParamNames}{The names of the parameters used in the expression. Example for generalised pareto: c("alpha", "sigma")}

\item{ParamEstimates}{Estimates obtained for the parameters. Following ordering of ParamNames}
}
\value{
Returns confidence bands for each of the variables, and the variance-covariance matrix
}
\description{
Calculates confidence bounds for pseudo-maximum likelihood using sandwich variance-covariance matrix (Geyer (2013): https://www.stat.umn.edu/geyer/5601/notes/sand.pdf).
}
\examples{
SandwichMaker(x=Value, w=Weights, EXPRESS=expression(w(log(1/sigma)-((1+alpha)/alpha)*log((sigma+alpha*(x-gamma.gep.hat1))/sigma))), ParamNames=c("alpha", "sigma"),ParamEstimates=c(1.8,1.2))
}
