% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPfunction.R
\name{TPfunction}
\alias{TPfunction}
\title{Estimates truncated type 1 pareto parameters}
\usage{
TPfunction(x, w = 1, gamma = NULL, nu = NULL, OptimSearch = "Default")
}
\arguments{
\item{x}{Data values used in parameter estimation.}

\item{w}{Data weights used in estimation. Default = 1.}

\item{gamma}{Sets the lower threshold parameter for the type 1 pareto. Default is NULL, setting the threshold to the minimum of x.}

\item{nu}{Sets the upper threshold parameter for the truncated type 1 pareto. Default is NULL, setting the threshold to the minimum of x.}

\item{OptimSearch}{The algorithm used in the optim function. If "Global", uses SANN in optim function. If "Local", uses L-BFGS-B in optim function. If "Default", uses the form from Zwijnenburg, Grilli & Engelbrecht (2022).}
}
\value{
Returns parameter estimates for alpha and sigma (and the minimum and maximum for x) for the truncated type 1 pareto.
}
\description{
Returns estimated parameters following the truncated type 1 pareto form.
}
\examples{
TPfunction(Value,Weights,min(Value),max(Value),OptimSearch="Default")
}
